+++ service/api/removeDislikeFromCommentRelatedToPost.go
@@ -20,8 +20,23 @@ func (rt *_router) removeDislikeFromCommentRelatedToPost(w http.ResponseWriter, 
 	var err error
 
 	var requestEmotionToComment RequestEmotionToComment
+	bearerToken := r.Header.Get("Bearer")
+	uid, errAuth := rt.db.AuthUid(bearerToken)
+	if errAuth != nil {
+		ctx.Logger.WithError(errAuth).Error("not authorized request")
+		w.WriteHeader(http.StatusUnauthorized)
+		return
+
+	}
 
 	json.NewDecoder(r.Body).Decode(&requestEmotionToComment)
+
+	if requestEmotionToComment.IdUser != uid {
+		ctx.Logger.WithError(errAuth).Error("not authorized request")
+		w.WriteHeader(http.StatusForbidden)
+		return
+
+	}
 
 	err = rt.db.RemoveDislikeFromCommentRelatedToPost(requestEmotionToComment.IdPost,
 		requestEmotionToComment.IdCommentEmotion, requestEmotionToComment.IdUser)
@@ -32,5 +47,6 @@ func (rt *_router) removeDislikeFromCommentRelatedToPost(w http.ResponseWriter, 
 		return
 	}
 	w.Header().Set("Content-Type", "application/json")
+	w.WriteHeader(http.StatusNoContent)
 
 }
